Add support for common variables.
Lump multiple lexemes into tokens for common variables.

I know that the official lexer creates something called a "set token"
that is a token that contains multiple child tokens and then the common
variable is a set token that consists of an article keyword and an
identifier.

Not sure this is the way to go. I feel like the classification of tokens
into simple/common might be better handled by syntax analyzer.